requirements.txt
asgiref==3.7.2
Django==5.0.2
django-quill-editor==0.1.40
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
